Started work on the editor
It is possible to move objects around and scroll through the stage.
The actual objects are still hard-coded, so I still have to add
functionality for selecting and placing them.
Also, the background is not drawn atm. Need to add functions for
changing its colour, as well as the size (well, width) of the stage.
The camera is currently stuck in the positive space (which is good), but
as stage width has yet to be implemented, it is possible to scroll
infinitely to the right.
I also did a little experiment in parallax scrolling, since that will be
in a (much) later version of the actual game.